deactivate czi to ome conversion (#41)
* deactivate czi to ome conversion

issue with the conversion in the pyramid recognition by FiJi!
Need to be patch.
in the mean time, deactivation

* to satisfy lint

* Add link to config.py

For the CZI-OME conversion, roll back the change in the code, was correct!
Instead, add the variable so that it gets them from the config.py (which is on the server and modifiable).
Use default value (current) if it does not found config.py